let eraseFromParent return iterator
